STY: updated casting
Updated casting to use int64 and float64.

+++ pysatSpaceWeather/instruments/methods/kp_ap.py
@@ -535,7 +535,7 @@ def combine_kp(standard_inst=None, recent_inst=None, forecast_inst=None,
     kp_inst.inst_module = pysat_sw.instruments.sw_kp
     kp_inst.tag = tag
     kp_inst.date = start
-    kp_inst.doy = int(start.strftime("%j"))
+    kp_inst.doy = np.int64(start.strftime("%j"))
     kp_inst.meta = pysat.Meta(labels=kp_inst.meta_labels)
     initialize_kp_metadata(kp_inst.meta, 'Kp', fill_val=fill_val)
